Initial Commit
First commit for dissertation files. This is the latest version, with
styles and basic skeleton copied from Dorathea Brosius' 2015 template.

By typing "make", the dissertation is made. All .aux files are kept and
left in their own ./aux/ directory for neatness. Currently, only one
of the styles is linked to from the local folder, while the rest are in the TeX directory on my machine.